services/credit/requirements.txt
flask==3.0.3
flask-cors==4.0.1
strawberry-graphql[flask]
supabase==2.28.3
requests==2.31.0
pika==1.3.2
python-dotenv==1.0.1
